Finished the rest of the project
Used the previous button I was messing with, and followed instructions on Odin Project to create a Rock Paper Scissors game. Uses constants to keep track of what the user or computer is playing, and debugged with console.log in order to make sure everything worked. Mostly used if else statements, but may impliment switch statements if possible. Utilized the messaging to set up computer score, human score, and tie counter. Also used CSS to format it a little nicer.